requirements.txt
fastapi
uvicorn
transformers
pillow
torch
google-images-search